pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hdc-nids"
version = "0.1.0"
description = "Drift-adaptive hyperdimensional NIDS research pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "matplotlib>=3.10",
  "numpy>=2.4",
  "PyYAML>=6.0",
  "scikit-learn>=1.8",
  "torch>=2.11",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
